assigning feat_stride to base_size
If we are putting the RPN/RoI layer on top of a feature map with another resolution rather than 1/16, we need to change the base_size too so it is consistent with the feat_stride and thus less confusing. Or maybe another config flag for base_size might be cleaner.